ia-topnav: use local ia-wayback-search package
Previously, the @internetarchive/ia-wayback-search package was used,
which a *published* package on comes from npmjs.org

This meant that the current functionality depended on old (previous
release), instead of the current, in-tree version of that module.

This further meant that in order for a local developer to work on the
wayback search, he would have to toggle the "import WaybackSearch from
'../../../packages/ia-wayback-search/index.js'" line in
wayback-search.js in order to avoid picking up the released module.

Since this repo is the source of truth (ie, contains the authoritative
source of ia-wayback-search), it is appropriate to use it as the source
of that module, and not whatever old version happens to be currently
published.

It is also not clear to me, why a released version was proposed in the
"future way we could move to" comment. It seems more elegant, but
innapropriate to treat a genuinely local module as an external one.

This commit removes the need for devs to do the "toggle" dance.