Remove yarn.lock to fix "invalid version" error
When trying to "npm uninstall @internetarchive/ia-wayback-search",
NPM complained that some version was invalid.

The error was not useful to point out what version was the problem,
and per https://stackoverflow.com/questions/71383116/npm-err-invalid-version-on-npm-install,
the solution was to remove the packages.lock file

This worked. I don't know what NPM is complaining about,
but the solution works, without apparent ill side-effects, so I'll take it.