Remove dependency on offshoot.prod.archive.org
Instead of using "lit" from archive.org, add it to the project as a normal dependency.

The offshot.prod.archive.org seems firewalled, so my local development efforts failed to correctly load the built website.

Reliance on an externally hosted module seems misguided anyway. Importing it as a module makes versioning it reliable.